+++ Surrounded/Assets/Scripts/BallShooter.cs
@@ -1,4 +1,4 @@
-using System.Collections;
+﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
@@ -9,7 +9,7 @@ public class BallShooter : MonoBehaviour {
     Transform trans;
 
     public GameObject ball;
-    float ballStartDistance = 0.15f; // space between player and ball
+    readonly float ballSpawnSpacing = 0.2f; // space between player and ball
 
     float ballRadius;
     float playerRadius;
@@ -23,7 +23,7 @@ public class BallShooter : MonoBehaviour {
 	}
 	
 	void Update () {
-        if (Input.GetButtonDown("Fire") && !Game.isPaused()) {
+        if (Input.GetButtonDown("Fire") && !Game.IsPaused()) {
             Vector3 mousePos = cam.ScreenToWorldPoint(Input.mousePosition);
 
             Vector2 difference = new Vector2(mousePos.x, mousePos.y) - new Vector2(trans.position.x, trans.position.y);
@@ -31,9 +31,13 @@ public class BallShooter : MonoBehaviour {
             angle = Vector2.Angle(Vector2.up, difference) * sign * Mathf.Deg2Rad; // angle of ball to be shot in radians
             Vector3 direction = new Vector3(Mathf.Sin(angle), Mathf.Cos(angle));
 
-            float adjustedAngle = angle % (90 * Mathf.Deg2Rad); // adjust because we only want a value between 0 and 45 degrees
+            Debug.Log("start");
+            float adjustedAngle = Mathf.Abs(angle) % (90 * Mathf.Deg2Rad); // adjust because we only want a value between 0 and 45 degrees
+            Debug.Log(adjustedAngle);
             if (adjustedAngle > (45 * Mathf.Deg2Rad)) adjustedAngle = (90 * Mathf.Deg2Rad) - adjustedAngle;
-            float distance = ballRadius + (playerRadius / Mathf.Cos(adjustedAngle)) + ballStartDistance; // ball radius + formula to calculate distance to edge of square + ballStartDistance (for spacing)
+            Debug.Log(adjustedAngle);
+            float distance = ballRadius + (playerRadius / Mathf.Cos(adjustedAngle)) + ballSpawnSpacing; // (ball radius) + (formula to calculate distance to edge of square) + (ballSpawnSpacing)
+            Debug.Log(distance);
 
             Vector3 ballPos = trans.position + (direction * distance);
             GameObject spawned = Instantiate(ball, ballPos, Quaternion.identity);