+++ Surrounded/Assets/Scripts/PauseMenu.cs
@@ -1,45 +1,46 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.UI;
+using TMPro;
 
 public class PauseMenu : MonoBehaviour {
 
     public GameObject pauseMenuUI;
 
-	// Use this for initialization
-	void Start () {
-		
-	}
+    public KeepScore score;
 	
 	// Update is called once per frame
 	void Update() {
         if (Input.GetKeyDown(KeyCode.Escape)) {
-            if (Game.isPaused()) Resume();
+            if (Game.IsPaused()) Resume();
             else Pause();
         }
 	}
 
     public void GoToMainMenu() {
+        Game.Instance.CheckAndSetScore(score.score);
+        Game.SetPaused(false);
+        Time.timeScale = 1f;
+        SaveLoad.Save();
         Game.LoadMainMenu();
     }
 
     public void Resume() {
         Camera.main.GetComponent<GameAudio>().Resume();
-        Game.setPaused(false);
+        Game.SetPaused(false);
         pauseMenuUI.SetActive(false);
         Time.timeScale = 1f;
     }
 
     public void Pause () {
         Camera.main.GetComponent<GameAudio>().Pause();
-        Game.setPaused(true);
+        Game.SetPaused(true);
         pauseMenuUI.SetActive(true);
         Time.timeScale = 0f;
     }
 
     public void SaveAndQuit() {
-        Game.Instance.CheckAndSetScore(FindObjectOfType<Text>().GetComponent<KeepScore>().score);
+        Game.Instance.CheckAndSetScore(score.score);
         SaveLoad.Save();
         Game.QuitGame();
     }