+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 0.6
+### Added
+- Two-hit mode (balls take two-hits to be destroyed instead of one)
+- Teleport mode (instead of bouncing, balls move off one side of the screen and reappear on the opposite side)
+- Highscores screen (shows highscores in all game modes)
+
+### Changed
+- Made menus and game scale to work on all window sizes
+
+### Fixed
+- Issue that can cause a player to be killed by a ball they just shot
+- Balls don’t correctly bounce off sides of screen after screen has been resized
+- Balls are white for one frame before changing to the correct colour
+- Options menu correctly saves quality, window size, and fullscreen settings
+
 ## 0.5.1 - 2019-01-28
 ### Added
 - Game icon
@@ -30,7 +45,7 @@
 - Improved positioning of text in menu
 - Centered text in options menu
 
-## 0.2.0 - 2018-07-17
+## 0.2 - 2018-07-17
 ### Added
 - A main menu (Play, Options, Credits, Save & Quit)
 - Game over screen